pyproject.toml
[tool.poetry]
name = "natal"
version = "0.1.0"
description = "create Natal Chart in SVG with ease"
license = "MIT"
repository = "https://github.com/hoishing/natal"
documentation = "https://hoishing.github.io/natal"
authors = ["Kelvin Ng <hoishing@gmail.com>"]
readme = "README.md"
packages = [{ include = "natal", from = "." }]

[tool.poetry.dependencies]
python = "^3.12"
pyswisseph = "^2.10.3.2"
ptag = "^0.1.1"
pydantic = "^2.8.2"
pyyaml = "^6.0.2"
pandas = "^2.2.2"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
mkdocs-material = "^9.5.31"
mkdocstrings = { extras = ["python"], version = "^0.25.2" }
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
